+++ previous-versions/common/codatcommon.go
@@ -164,9 +164,9 @@ func New(opts ...SDKOption) *CodatCommon {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "3.0.0",
-			SDKVersion:        "0.29.0",
-			GenVersion:        "2.188.1",
-			UserAgent:         "speakeasy-sdk/go 0.29.0 2.188.1 3.0.0 github.com/codatio/client-sdk-go/previous-versions/common",
+			SDKVersion:        "0.29.1",
+			GenVersion:        "2.194.1",
+			UserAgent:         "speakeasy-sdk/go 0.29.1 2.194.1 3.0.0 github.com/codatio/client-sdk-go/previous-versions/common",
 		},
 	}
 	for _, opt := range opts {