+++ previous-versions/common/RELEASES.md
@@ -383,3 +383,13 @@ Based on:
 - [go v0.29.0] previous-versions/common
 ### Releases
 - [Go v0.29.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/common/v0.29.0 - previous-versions/common
+
+## 2023-11-20 01:27:49
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.29.1] previous-versions/common
+### Releases
+- [Go v0.29.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/common/v0.29.1 - previous-versions/common